Added util file for all columns and hardcoded rows
