+++ cradlepoint/pages/2home.jsx
@@ -6,6 +6,7 @@ import CPButton from '../components/button/CPButton';
 import CreateNewModal from './createNewModal';
 import NewEngagModalInfo from './newEngagModalInfo';
 import NewModalClone from './newModalClone';
+import { engagementColumns, engagementRows } from '../util/tableColumns';
 
 // TODO: adjust scaling and font of the page
 export default function HomeScreen(props) {
@@ -27,16 +28,7 @@ export default function HomeScreen(props) {
     
     const classes = useStyles();
 
-    const engagementColumns = [
-        { field: 'id', headerName: 'id', headerClassName: 'header', flex: 1, hide: true},
-        { field: 'name', headerName: 'Name', headerClassName: 'header', flex: 1},
-        { field: 'status', headerName: 'Status', headerClassName: 'header', flex: 1},
-        { field: 'details', headerName: 'Details', headerClassName: 'header', sortable:false, flex: 2, hide: true, minWidth: 200},
-        { field: 'sysEng', headerName: 'SEng', headerClassName: 'header', flex: 1},
-        { field: 'pocEng', headerName: 'POC Eng', headerClassName: 'header', flex: 1},
-        { field: 'customer', headerName: 'Customer', headerClassName: 'header', flex: 1},
-        { field: 'sfdc', headerName: 'SFDC', headerClassName: 'header', flex: 1},
-        { field: 'dateCreated', headerName: 'Date Created', headerClassName: 'header', flex: 1},
+    const engagementColumnsWithActions = engagementColumns.concat([
         { 
           field: 'button', 
           flex: 1,
@@ -48,17 +40,7 @@ export default function HomeScreen(props) {
             <CPButton text="DETAILS"/>
           )
         }
-      ];
-
-// TODO: make rows not hard coded (delete later)
-    const rows = [
-    {id: 1, name: 'Engagement 1', status: 'Pending', details: ' ', sysEng: 'John Rogers',	pocEng: 'Paul Switchport', customer: 'ABC Bus Company', sfdc: 'https://cradlepoint.lightning.force.com/lightning/r/Opportunity/0063800000qtILXAA2/view', dateCreated: '10/01/2021'},
-    {id: 2, name: 'Engagement 2', status: 'Assigned', details: ' ', sysEng: 'Michael Smith', pocEng: 'George Packets', customer: 'Big Finance', sfdc: 'https://cradlepoint.lightning.force.com/lightning/r/Opportunity/0063800000qtILXAA2/view', dateCreated: '10/02/2021'},
-    {id: 3, name: 'Engagement 3', status: 'POC testing complete', details: ' ', sysEng: 'Don Lee', pocEng: 'Ron State', customer: 'SensorCo', sfdc: 'https://cradlepoint.lightning.force.com/lightning/r/Opportunity/0063800000qtILXAA2/view', dateCreated: '10/03/2021'},
-    {id: 4, name: 'Engagement 4', status: 'POC testing outcome', details: ' ', sysEng: 'Jim Black', pocEng: 'Jason Dumps', customer: 'Burgerz-R-us', sfdc: 'https://cradlepoint.lightning.force.com/lightning/r/Opportunity/0063800000qtILXAA2/view', dateCreated: '10/04/20201'},
-    {id: 5, name: 'Engagement 5', status: 'POC approved', details: ' ', sysEng: 'Don Lee', pocEng: 'Paul Switchport', customer: 'SensorCo', sfdc: 'https://cradlepoint.lightning.force.com/lightning/r/Opportunity/0063800000qtILXAA2/view', dateCreated: '10/05/20201'},
-    {id: 6, name: 'Engagement 6', status: 'Archieved', details: ' ', sysEng: 'Jim Black', pocEng: 'George Packets', customer: 'Burgerz-R-us', sfdc: 'https://cradlepoint.lightning.force.com/lightning/r/Opportunity/0063800000qtILXAA2/view', dateCreated: '10/06/20201'}
-    ]
+      ]);
 
     const [selectModalOpen, setSelectModalOpen] = useState(false);
     const [infoModalOpen, setInfoModalOpen] = useState(false);
@@ -68,10 +50,6 @@ export default function HomeScreen(props) {
   
     
     function updateModal(modalType,rowId=0){
-      // console.log(input);
-      // setModalType(input);
-      console.log("modalOpen called with ");
-      console.log(modalType);
       if (modalType === "select"){
         setSelectModalOpen(true)
       } else if (modalType === "scratch"){
@@ -102,16 +80,17 @@ export default function HomeScreen(props) {
                 }
               }
             />
-            <PlainTable rows={rows} columns={engagementColumns} className={classes.root}/>
+            {/* TODO: get rows from database */}
+            <PlainTable rows={engagementRows} columns={engagementColumnsWithActions} className={classes.root}/>
             <CreateNewModal
-              testPlanOrEngagement={"ENGAGEMENT"}
+              type={"Engagement"}
               modalOpen={selectModalOpen} 
               onClickNext={updateModal}
               onClose={()=> setSelectModalOpen(false)}
             />
 
             <NewEngagModalInfo
-              testPlanOrEngagement={"ENGAGEMENT"}
+              type={"Engagement"}
               modalOpen={infoModalOpen} 
               onBack={()=> setInfoModalOpen(false)}
               selectedRow={selectedRow}
@@ -119,7 +98,7 @@ export default function HomeScreen(props) {
               />
 
             <NewModalClone
-              testPlanOrEngagement={"ENGAGEMENT"}
+              type={"Engagement"}
               modalOpen={cloneModalOpen} 
               onClickNext={updateModal}
               onBack={()=> setCloneModalOpen(false)}
