+++ cradlepoint/pages/createNewModal.js
@@ -7,10 +7,10 @@ import PropTypes from 'prop-types';
 export default function CreateNewModal(props) {
   return (
       <Modal className={styles.Modal} isOpen={props.modalOpen}>
-        <h2>Create New {props.isEngagementOrTestPlan === "TEST_PLAN" ? "Test Plan" : "Engagement"}</h2>
+        <h2>Create New {props.type}</h2>
         <CPButton text='From scratch' className="ModalButton" onClick={()=>{console.log('scratch clicked');
         props.onClickNext("scratch")}}/> 
-        <CPButton text='From exisiting engagement (Clone)' onClick={()=>props.onClickNext("clone")}/>
+        <CPButton text={'From exisiting ' + props.type + ' (Clone)'} onClick={()=>props.onClickNext("clone")}/>
         <CPButton text='Cancel' onClick={props.onClose}/>
       </Modal>
   );
@@ -20,6 +20,6 @@ CreateNewModal.propTypes = {
   onClose: PropTypes.bool.isRequired,
   modalOpen: PropTypes.bool.isRequired,
   onClickNext:PropTypes.func.isRequired,
-  testPlanOrEngagement: PropTypes.string.isRequired,
+  type: PropTypes.string.isRequired,
 }
