+++ cradlepoint/pages/deviceModals/selectQuantity.jsx
@@ -8,6 +8,7 @@ import Checkbox from '@mui/material/Checkbox';
 import { pink } from '@mui/material/colors';
 
 import { Field, Formik} from 'formik';
+import { BOMColumns, BOMRows, LibraryBOMColumns } from "../../util/tableColumns";
 
 export default function SelectQuantity(props) {
   const useStyles = makeStyles({
@@ -27,11 +28,7 @@ export default function SelectQuantity(props) {
 
 const classes = useStyles();
 
-const deviceColumns = [
-    { field: 'id', headerName: 'id', headerClassName: 'header', flex: 1, hide: true},
-    { field: 'deivceName', headerName: 'Device Name', headerClassName: 'header', flex: 1},
-    { field: 'physical', headerName: 'Physical', headerClassName: 'header', flex: 1},
-    { field: 'sku', headerName: 'SKU', headerClassName: 'header', flex: 1},
+const BOMColumnsWithFields = LibraryBOMColumns.concat([
     { 
       field: 'quantity', 
       flex: 1,
@@ -74,25 +71,15 @@ const deviceColumns = [
                 />
             )
         }
-    }
-  ];
+    }]);
 
 
-    // TODO: make rows not hard coded (delete later)
-    const rows = [
-        {id: 1, deivceName: 'Device 1', sku: 'this is the sku'},
-        {id: 2, deivceName: 'Device 2', sku: 'this is the sku'},
-        {id: 3, deivceName: 'Device 3', sku: 'this is the sku'},
-        {id: 4, deivceName: 'Device 4', sku: 'this is the sku'},
-        {id: 5, deivceName: 'Device 5', sku: 'this is the sku'},
-        {id: 6, deivceName: 'Device 6', sku: 'this is the sku'}
-    ]
 
   return (
     <>
       <Modal className={styles.Modal} isOpen={props.modalOpen}>
         <h2>Add new device(s) to the summary BOM</h2>
-        <PlainTable rows={rows} columns={deviceColumns} className={classes.root} 
+        <PlainTable rows={BOMRows} columns={BOMColumnsWithFields} className={classes.root} 
         />
         <CPButton text='Back' onClick={props.onBack}/>
         <CPButton text='Add' />
